moved shell back to util
