+++ benchmark/job/cli.py
@@ -2,7 +2,6 @@ import logging
 from typing import Optional
 
 import click
-from gymlib.shell import subprocess_run
 from gymlib.symlinks_paths import (
     get_tables_dirname,
     get_workload_dirname,
@@ -12,6 +11,7 @@ from gymlib.symlinks_paths import (
 from gymlib.workspace import DBGymWorkspace, fully_resolve_path
 
 from benchmark.constants import DEFAULT_SCALE_FACTOR
+from util.shell import subprocess_run
 
 JOB_TABLES_URL = "https://event.cwi.nl/da/job/imdb.tgz"
 JOB_QUERIES_URL = "https://event.cwi.nl/da/job/job.tgz"
