+++ dbms/postgres/cli.py
@@ -11,7 +11,6 @@ from typing import Any, Optional
 import click
 import sqlalchemy
 from gymlib.pg import create_sqlalchemy_conn, sql_file_execute
-from gymlib.shell import subprocess_run
 from gymlib.symlinks_paths import (
     get_dbdata_tgz_symlink_path,
     get_pgbin_symlink_path,
@@ -32,6 +31,7 @@ from benchmark.constants import DEFAULT_SCALE_FACTOR
 from benchmark.job.load_info import JobLoadInfo
 from benchmark.tpch.load_info import TpchLoadInfo
 from dbms.load_info_base_class import LoadInfoBaseClass
+from util.shell import subprocess_run
 
 DBGYM_POSTGRES_USER = "dbgym_user"
 DBGYM_POSTGRES_PASS = "dbgym_pass"
